Move user-related presenters to own folder
